requirements.txt
# FastAPI Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19  # For file uploads

# Gemini Vision API
google-generativeai==0.8.3

# Image Processing
Pillow==11.0.0

# Environment Variables
python-dotenv==1.0.1

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1  # For testing FastAPI endpoints
